Re-run MFoR analysis with masked RectBivariateSpline interpolation
A single contour path per specified velocity deficit level is _not_
guaranteed. Selection based on maximum path length is unreliable;
instead, assume that the "correct" contour is that which is closest to
the origin (in the MFoR). This is selected by calculating the geometric
center of the contour and verifying that it is within 1D of the origin
(an ad hoc value).