Update processing of 20Hz data
Create dataframe for each group so that the height index level gets
properly translated into xarray.Dataset coordinates. Previously, the
first xarray.Dataset DataArray assignment sets (and fixes) the
coordinates, so that additional assignments for other output channels
for the same sensor group are not able to modify the coordinates.